requirements.txt
requests
pandas
numpy
scikit-learn
python-dotenv
pyyaml
pytz
joblib
